Extracted response validation
