+++ src/log4stash/ElasticClient/IResponseValidator.cs
@@ -1,0 +1,9 @@
+﻿using RestSharp;
+
+namespace log4stash.ElasticClient
+{
+    public interface IResponseValidator
+    {
+        void ValidateResponse(IRestResponse response);
+    }
+}
