+++ src/log4stash/log4stash.csproj
@@ -79,9 +79,11 @@
     <Compile Include="ElasticClient\IElasticsearchClient.cs" />
     <Compile Include="ElasticClient\InnerBulkOperation.cs" />
     <Compile Include="ElasticClient\IRequestFactory.cs" />
+    <Compile Include="ElasticClient\IResponseValidator.cs" />
     <Compile Include="ElasticClient\PartialElasticResponse.cs" />
     <Compile Include="ElasticClient\RequestDetails.cs" />
     <Compile Include="ElasticClient\RequestFactory.cs" />
+    <Compile Include="ElasticClient\ResponseValidator.cs" />
     <Compile Include="ElasticClient\RestSharp\IRestClientFactory.cs" />
     <Compile Include="ElasticClient\RestSharp\RestSharpClientFactory.cs" />
     <Compile Include="ElasticClient\WebElasticClientFactory.cs" />
